Router and testing files
Added router.js with the routing logic. Changed the imdex.html so that it uses id app, and added pages/students.js for testing.
Does not work if we refresh the page that is not the landing page yet.
The "home" is for now a variable in the router.js file.